pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyqual"
version = "0.1.140"
description = "Declarative quality gate loops for AI-assisted development"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [{name = "Tom Sapletta", email = "tom@sapletta.com"}]
keywords = ["AI", "CI/CD", "quality-gates", "LLM", "pipeline", "devtools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "litellm>=1.0",
    "python-dotenv>=1.0",
    "nfo>=0.2.13",
]

[project.optional-dependencies]
analysis = ["code2llm", "vallm", "redup"]
costs = ["costs"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "goal>=2.1.0", "costs>=0.1.20", "pfix>=0.1.60", "tox>=4.0.0"]
llx = ["llx>=0.1.47"]
mcp = ["llx>=0.1.47", "mcp>=1.0", "uvicorn>=0.29"]
all = ["llx>=0.1.47", "code2llm", "vallm", "redup", "costs"]

[project.scripts]
pyqual = "pyqual.cli:app"

[tool.setuptools]
packages = ["pyqual", "pyqual.integrations"]

[tool.setuptools.package-data]
pyqual = ["default_tools.json"]

[tool.costs]
model = "openrouter/x-ai/grok-code-fast-1"
all_commits = true

[tool.pfix]
# Self-healing Python configuration
model = "openrouter/x-ai/grok-code-fast-1"
auto_apply = true
auto_install_deps = true
auto_restart = false
max_retries = 3
create_backups = false
git_auto_commit = false

[tool.pfix.runtime_todo]
enabled = true
todo_file = "TODO.md"
min_severity = "low"
deduplicate = true

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py39,py310,py311,py312
isolated_build = true

[testenv]
deps =
    pytest>=8.0
    pytest-cov>=5.0
    pfix>=0.1.60
extras = all
commands = pytest {posargs}

[testenv:lint]
deps = ruff
commands = ruff check pyqual/

[testenv:pyqual]
deps =
    pytest>=8.0
    pytest-cov>=5.0
commands = pyqual run --config pyqual.yaml --workdir .
"""
